requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
requests==2.31.0
elevenlabs==0.2.26
pydub==0.25.1
Werkzeug==2.3.7
gunicorn==21.2.0
python-multipart==0.0.6
Pillow==10.1.0
WTForms==3.1.1
email-validator==2.1.0
cryptography==41.0.7
bcrypt==4.1.2
redis==5.0.1
celery==5.3.4 